Austin/integrate xtb opt (#17)
* Implement xtb submission file generation

Files for high-throughput optimization via xtb are generated. This is
accomplished in the working generation script. It would be ideal to work
this feature into the larger package in the future.

* Add xtb structure generation functionality

Molecule can now generate the starting structure for the xtb geometry
optimization step.

* Allow user to write_xyz

In addition to reading xyz files, SMORES can now write xyz files